Trying to clean up the workout contec:\Users\Jesse\programming-projects\workout-app\src\contexts\WorkoutContext.js
