+++ src/components/Workout/WorkoutSelection.js
@@ -1,17 +1,12 @@
 import React from 'react'
 import SingleWorkout from './SingleWorkout'
 import { e1 } from '../../assets/data/e1'
-
-const WorkoutSelection = ({ startWorkout }) => {
-  return (
-    <div>
-      <SingleWorkout
-        text={'Your Next Workout'}
-        exercise={e1}
-        startWorkout={startWorkout}
-      />
-    </div>
-  )
+const WorkoutSelection = () => {
+    return (
+        <div>
+            <SingleWorkout text={'Your Next Workout'} exercise={e1} />
+        </div>
+    )
 }
 
 export default WorkoutSelection
